split the gtm / aep from the core tms
