+++ feature-libs/tracking/tms/aep/config/default-aep.config.ts
@@ -1,5 +1,5 @@
 import { TmsConfig } from '@spartacus/tracking/tms/core';
-import { AepCollectorService } from '../aep-collector.service';
+import { AepCollectorService } from '../services/aep-collector.service';
 
 export const defaultAdobeExperiencePlatformConfig: TmsConfig = {
   tagManager: {
