+++ feature-libs/tracking/tms/aep/ng-package.json
@@ -2,6 +2,8 @@
   "$schema": "../../../../node_modules/ng-packagr/ng-package.schema.json",
   "lib": {
     "entryFile": "./public_api.ts",
-    "umdModuleIds": {}
+    "umdModuleIds": {
+      "@spartacus/core": "core"
+    }
   }
 }
