+++ feature-libs/tracking/tms/aep/public_api.ts
@@ -1,2 +1,3 @@
-export * from './aep-collector.service';
+export * from './aep.module';
 export * from './config/index';
+export * from './services/index';
