+++ feature-libs/tracking/tms/aep/services/index.ts
@@ -1,0 +1,1 @@
+export * from './aep-collector.service';
