+++ feature-libs/tracking/tms/gtm/config/default-gtm.config.ts
@@ -1,5 +1,5 @@
 import { TmsConfig } from '@spartacus/tracking/tms/core';
-import { GtmCollectorService } from '../gtm-collector.service';
+import { GtmCollectorService } from '../services/gtm-collector.service';
 
 export const defaultGoogleTagManagerConfig: TmsConfig = {
   tagManager: {
