+++ feature-libs/tracking/tms/gtm/gtm.module.ts
@@ -1,0 +1,13 @@
+import { ModuleWithProviders, NgModule } from '@angular/core';
+import { provideDefaultConfig } from '@spartacus/core';
+import { defaultGoogleTagManagerConfig } from './config/default-gtm.config';
+
+@NgModule({})
+export class GtmModule {
+  static forRoot(): ModuleWithProviders<GtmModule> {
+    return {
+      ngModule: GtmModule,
+      providers: [provideDefaultConfig(defaultGoogleTagManagerConfig)],
+    };
+  }
+}
