+++ feature-libs/tracking/tms/gtm/public_api.ts
@@ -1,2 +1,3 @@
 export * from './config/index';
-export * from './gtm-collector.service';
+export * from './gtm.module';
+export * from './services/index';
