+++ feature-libs/tracking/tms/gtm/services/index.ts
@@ -1,0 +1,1 @@
+export * from './gtm-collector.service';
